Fix bug causing the last episode of gap-free files not to be read
The number of samples read from the last episode of a gap-free file by
ABF2_ReadChannel() or ABF2_MultiplexRead() is determined by
CFileDescriptor::m_uLastEpiSize. A comment in ABF2_GetSynchEntry()
indicates that the last episode size should be set in the ABF file open
function. This is done in ABF2_ReadOpen() (via ReadEDVarLenSynch() which
in turn calls _SetChunkSize()), but libstfio implements its own open
function which does not do this. As a result, the default value of zero is
always returned, such that no samples are read from the last episode.

The only exported function that can set CFileDescriptor::m_uLastEpiSize is
ABF2_SetChunkSize(), so this commit adds a call to this function in
stfio::importABF2File(), using the current episode size as the chunk size
so that it sets the last episode size but does not change the number of
samples per episode.